make `Schema.scala` more independent
Move the materialization methods from the `Schema` object into a
new `SchemaMaterializer` object in the main module.

Move the `Violation` trait from the main module into the new
`schema` module, along with some subtypes required by the
`schema` code.
